+++ demos/manual/src/main/java/is/codion/framework/demos/manual/keybinding/KeyBindingModel.java
@@ -76,7 +76,7 @@ final class KeyBindingModel {
 	}
 
 	private void bindEvents(FilteredComboBoxModel<Item<LookAndFeelProvider>> lookAndFeelComboBoxModel) {
-		componentComboBoxModel.refresher().addRefreshListener(tableModel::refresh);
+		componentComboBoxModel.refresher().refreshObserver().addListener(tableModel::refresh);
 		componentComboBoxModel.addSelectionListener(component -> tableModel.refresh());
 		lookAndFeelComboBoxModel.addSelectionListener(lookAndFeelProvider -> componentComboBoxModel.refresh());
 	}