+++ swing/common-ui-tools/src/main/java/is/codion/swing/common/ui/tools/loadtest/LoadTestPanel.java
@@ -112,7 +112,7 @@ public final class LoadTestPanel<T> extends JPanel {
 	private LoadTestPanel(LoadTestModel<T> loadTestModel) {
 		this.loadTestModel = requireNonNull(loadTestModel, "loadTestModel");
 		this.loadTest = loadTestModel.loadTest();
-		this.loadTestModel.applicationTableModel().refresher().addRefreshFailedListener(this::displayException);
+		this.loadTestModel.applicationTableModel().refresher().refreshFailedObserver().addDataListener(this::displayException);
 		initializeUI();
 	}