+++ changelog.md
@@ -6,6 +6,7 @@ Codion Change Log
 - TableSelectionModel, add/removeListener methods replaced with methods that return the underlying event observer.
 - ColumnConditionModel, add/removeListener methods replaced with methods that return the underlying event observer.
 - TableConditionModel, add/removeListener methods replaced with methods that return the underlying event observer.
+- FilteredModel.Refresher, add/removeListener methods replaced with methods that return the underlying event observer.
 ### is.codion.framework.model
 - EntityEditModel.setDefaults() renamed defaults().
 - EntityEditModel.setDefault() renamed defaultValue(), now Value based.